+++ docs/HISTORY.rst
@@ -4,8 +4,9 @@ There's a frood who really knows where his towel is
 1.0a4 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Add banner tile that will replace image and link tiles. (closes `#218`_).
-  [hvelarde]
+- Add banner tile that will replace image and link tiles; add deprecation
+  warning to image and link tiles as they will be removed from package on
+  next release. (closes `#218`_). [hvelarde]
 
 - Refactor image tile to use original image and scales, when possible.
   [ericof, hvelarde]